Refactor namespaces and update DTOs for consistency
Changed namespaces from `Domain.Dtos` to `Business.Dtos` for better organization of data transfer objects. Updated `EditClientForm` to use a string for the `Id` property and added `ClientImage` for image uploads. Marked `UserId` in `EditUserProfileForm` as required. Enhanced validation in `UserSignInForm` and `UserSignUpForm`. Updated service interfaces and implementations to reflect the new namespace. Adjusted `_ViewImports.cshtml` to align with the changes.